requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0
chromadb>=0.5.0
pyyaml>=6.0
rich>=13.0
